requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
cryptography==39.0.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==3.2.15
django-allauth==0.52.0
django-bootstrap4==22.3
django-import-export==3.3.1
et-xmlfile==1.1.0
frozenlist==1.3.3
future==0.18.3
googlemaps==4.10.0
idna==3.4
importlib-metadata==2.1.3
line-bot-sdk==2.4.1
MarkupPy==1.14
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184278/work
mkl-service==2.4.0
multidict==6.0.4
numpy @ file:///C:/b/abs_datssh7cer/croot/numpy_and_numpy_base_1672336199388/work
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
pandas @ file:///C:/ci/pandas_1624383612827/work
Pillow==9.4.0
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
pypng==0.20220715.0
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-slugify==8.0.0
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0.1
qrcode==7.4.2
requests==2.28.1
requests-oauthlib==1.3.1
simpleui==4.0.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
slugify==0.0.1
soupsieve==2.3.2.post1
sqlparse==0.4.3
stripe==5.0.0
tablib==3.5.0
text-unidecode==1.3
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
whitenoise==6.3.0
wincertstore==0.2
xlrd==2.0.1
xlwt==1.3.0
yarl==1.8.2
zipp==3.11.0
